pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "retail-ops-mlops-lakehouse"
version = "0.1.0"
description = "Retail ops MLOps lakehouse learning project"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
  "numpy>=1.22.4",
  "pandas>=2.2",
  "pyarrow>=16",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "pydantic>=2.6",
  "rich>=13.7",
  "tqdm>=4.66",
  "typer>=0.12",
  "joblib>=1.3",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "jinja2>=3.1",
]

[project.optional-dependencies]
kaggle = ["kaggle>=1.7.4.5,<1.8"]
dev = [
  "ruff>=0.4",
  "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/retail_ops_mlops"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
